Storage: change shift logic
Assume a vertex, call it v, with degree d<7 is selected in the random process. Then:
d=1: one of the vertices with degree 9 is selected randomly as its neighbor.
d=2: two of the vertices with degree 9 are selected randomly as its neighbors.
d=3: all the vertices with degree 9 are the neighbors of v.
d=4: all the vertices with degree 9 and one random vertex with degree 8 are the neighbors of v.
d=5: all the vertices with degree 8 and 9 are the neighbors of v.
d=6: all the vertices with degree 8 and 9 and one random vertex with degree 7 are the neighbors of v.

Earlier, my understanding of the neighbour selection was:

d=4: among the vertices with degree 9 and 8 (5 vertices) 4 vertices should be selected randomly
d=5: all vertices with degree 8 and 9 are selected
d=6: among the vertices with degree 9, 8 and 7 (>= 7 vertices) 6 vertices should be selected randomly

Prof. Bar Noy says:
"Your implementation doesn't guarantee termination.
The HH works only if you connect the new vertex with degree d to d vertices with the current
largest degree breaking ties arbitrarily."