Storage: run selection process over vertices not degrees
D() is now the sum of degrees for all vertices (raised to q)
For example, in the sequence 2, 1, 1 and q = 1, D is 4
Earlier, D was set to be 2 + 1 (the sum of the different degrees)

weight() is now set for each vertex (earlier it was set for each degree)

Finally, since weights are computed for each vertex, the specific vertex
can be returned (earlier, a random vertex with the selected degree was
returned).

This fixes the issue where large negative q was choosing the full cycle
(it should be large positive q).